requirements.txt
#  Date: 18.01.2025
# 
#  Author: Amanda Uccello
#  Class: ICS4UR-1
#  School: Port Credit Secondary School
#  Teacher: Mrs. Kim
#  Description: 
#      This list all the python requirements needed to run this project

fastapi
logfire
pydantic_ai
opentelemetry-instrumentation-asgi
uvicorn
pydantic-ai>=0.0.12
python-dotenv>=1.0.1
psycopg2-binary
python-multipart>=0.0.20
httpx
pymupdf
weaviate-client
sentence-transformers
numpy<2